# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AppointmentConfirmationRead": {
        "additionalProperties": false,
        "properties": {
          "overlap_warnings": {
            "items": {
              "$ref": "#/components/schemas/ScheduleOverlapWarning"
            },
            "title": "Overlap Warnings",
            "type": "array"
          },
          "service_request": {
            "$ref": "#/components/schemas/ServiceRequestRead"
          }
        },
        "required": [
          "service_request",
          "overlap_warnings"
        ],
        "title": "AppointmentConfirmationRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntakeSettings": {
        "additionalProperties": false,
        "properties": {
          "horizon_days": {
            "default": 14,
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Horizon Days",
            "type": "integer"
          },
          "response_hours": {
            "default": 4,
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Response Hours",
            "type": "integer"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/IntakeSlot"
            },
            "maxItems": 12,
            "title": "Slots",
            "type": "array"
          },
          "urgencies": {
            "items": {
              "$ref": "#/components/schemas/UrgencyOption"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Urgencies",
            "type": "array"
          },
          "version": {
            "default": 1,
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          },
          "weekdays": {
            "items": {
              "maximum": 6.0,
              "minimum": 0.0,
              "type": "integer"
            },
            "maxItems": 7,
            "title": "Weekdays",
            "type": "array"
          }
        },
        "required": [
          "urgencies",
          "weekdays",
          "slots"
        ],
        "title": "IntakeSettings",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Coffix API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/service-intake-settings": {
      "get": {
        "operationId": "get_intake_settings_api_v1_admin_service_intake_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Intake Settings",
        "tags": [
          "service"
        ]
      },
      "put": {
        "operationId": "put_intake_settings_api_v1_admin_service_intake_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntakeSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Intake Settings",
        "tags": [
          "service"
        ]
      }
    },
...
    "/api/v1/admin/service-requests/{request_id}": {
      "get": {
        "operationId": "admin_service_detail_api_v1_admin_service_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Service Detail",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/appointment": {
      "post": {
        "operationId": "confirm_service_appointment_api_v1_admin_service_requests__request_id__appointment_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentConfirmation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentConfirmationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Service Appointment",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/appointment-preview": {
      "post": {
        "operationId": "preview_service_appointment_api_v1_admin_service_requests__request_id__appointment_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentConfirmation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleOverlapWarning"
                  },
                  "title": "Response Preview Service Appointment Api V1 Admin Service Requests  Request Id  Appointment Preview Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Service Appointment",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/assignment": {
      "post": {
        "operationId": "change_service_assignment_api_v1_admin_service_requests__request_id__assignment_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Service Assignment",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/cancel": {
      "post": {
        "operationId": "admin_service_cancel_api_v1_admin_service_requests__request_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCancelInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Service Cancel",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/diagnostic-fee": {
      "post": {
        "operationId": "set_diagnostic_fee_api_v1_admin_service_requests__request_id__diagnostic_fee_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticFeeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Diagnostic Fee",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/no-cost-repair": {
      "post": {
        "operationId": "start_no_cost_repair_api_v1_admin_service_requests__request_id__no_cost_repair_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start No Cost Repair",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/notes": {
      "post": {
        "operationId": "admin_service_note_api_v1_admin_service_requests__request_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminNoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceNoteRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Service Note",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/quote": {
      "post": {
        "operationId": "create_service_quote_api_v1_admin_service_requests__request_id__quote_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceQuoteCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Service Quote",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-requests/{request_id}/status": {
      "post": {
        "operationId": "update_admin_service_status_api_v1_admin_service_requests__request_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceOperationalAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Admin Service Status",
        "tags": [
          "service"
        ]
      }
    },
    "/api/v1/admin/service-types": {
      "get": {
        "operationId": "list_service_types_api_v1_admin_service_types_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceTypeRead"
                  },
                  "title": "Response List Service Types Api V1 Admin Service Types Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Service Types",
        "tags": [
          "service"
        ]
      },
      "post": {
        "operationId": "create_service_type_api_v1_admin_service_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceTypeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Service Type",
        "tags": [
          "service"
        ]
      }
    },
...
  }
}